resized components to fit on android device for testing
