+++ StyleLights/Screens/Customization/CustomizeColorActivity.cs
@@ -1,0 +1,39 @@
+﻿
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+
+using Android.App;
+using Android.Content;
+using Android.OS;
+using Android.Runtime;
+using Android.Views;
+using Android.Widget;
+
+namespace StyleLights
+{
+	[Activity (Label = "CustomizeColorActivity")]			
+	public class CustomizeColorActivity : Activity
+	{
+		protected override void OnCreate (Bundle bundle)
+		{
+			base.OnCreate (bundle);
+			SetContentView (Resource.Layout.CustomizeLighting);
+
+			var metrics = Resources.DisplayMetrics;
+			var widthInDp = ConvertPixelsToDp (metrics.WidthPixels);
+			var heightInDp = ConvertPixelsToDp (metrics.HeightPixels);
+
+
+
+
+		}
+
+		private int ConvertPixelsToDp(float pixels) {
+			var dp = (int) ((pixels) / Resources.DisplayMetrics.Density);
+			return dp;
+		}
+	}
+}
+
