--- StyleLights/Screens/Selection/CustomizeLightingActivity.cs
@@ -1,44 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-using Android.App;
-using Android.Content;
-using Android.OS;
-using Android.Runtime;
-using Android.Views;
-using Android.Widget;
-
-namespace StyleLights
-{
-	[Activity (Label = "CustomizeLightingActivity")]
-	public class CustomizeLightingActivity : Activity
-	{
-		protected override void OnCreate (Bundle bundle)
-		{
-			base.OnCreate (bundle);
-			SetContentView (Resource.Layout.CustomizeLighting);
-
-			var metrics = Resources.DisplayMetrics;
-			var widthInDp = ConvertPixelsToDp (metrics.WidthPixels);
-			var heightInDp = ConvertPixelsToDp (metrics.HeightPixels);
-
-			var selectionButton = FindViewById<Button> (Resource.Id.SelectionButton);
-
-			if (selectionButton != null) {
-				selectionButton.Click += (sender, e) => {
-					var activity = new Intent(this, typeof(MainActivity));
-					StartActivity(activity);
-				};
-			}
-
-		}
-
-		private int ConvertPixelsToDp(float pixels) {
-			var dp = (int) ((pixels) / Resources.DisplayMetrics.Density);
-			return dp;
-		}
-	}
-}
-
